refactor(robot-server): more little accesses
Annoyingly, there were a couple buried places we were synchronously
hitting the hardware controller:

- The estop handler was doing it a million times, which is bad since it
was polled
- There were some places using the protocol engine's method of
determining if this is an ot3 or not rather than the server's method of
determining if this is an ot3 or not; the server detects subprocess mode
and shortcircuits, the engine doesn't.

Also why were there so many feature flag mock fixtures
