+++ robot-server/robot_server/service/legacy/routers/motors.py
@@ -7,15 +7,13 @@ from starlette import status
 
 from opentrons.hardware_control import HardwareControlAPI
 from opentrons.hardware_control.types import Axis
-from opentrons.protocol_engine.errors import HardwareNotSupportedError
-from opentrons.protocol_engine.resources.ot3_validation import ensure_ot3_hardware
 from opentrons_shared_data.errors import ErrorCodes
 from server_utils.audit.fastapi import get_audit_logger
 from server_utils.auth.resource_server.fastapi import require_scopes
 from server_utils.auth.scopes import Scope
 
-from robot_server.errors.error_responses import LegacyErrorResponse
-from robot_server.hardware import get_hardware
+from robot_server.errors.error_responses import LegacyErrorResponse, ApiError
+from robot_server.hardware import get_hardware, get_ot3_hardware
 from robot_server.service.legacy.models import V1BasicResponse
 from robot_server.service.legacy.models import motors as model
 
@@ -62,8 +60,8 @@ async def post_disengage_motors(
 ) -> V1BasicResponse:
     input_axes = [Axis[ax.upper()] for ax in axes.axes]
     try:
-        hardware = ensure_ot3_hardware(hardware)
-    except HardwareNotSupportedError:
+        hardware = get_ot3_hardware(hardware)
+    except ApiError:
         # Filter out non-ot2 axes when running on OT2
         input_axes = [axis for axis in input_axes if axis in Axis.ot2_axes()]
